:sparkles: jwtProperties
