+++ notdo-infrastructure/src/main/kotlin/com/gsm/notdo/global/security/priciple/AuthDetails.kt
@@ -2,10 +2,8 @@ package com.gsm.notdo.global.security.priciple
 
 import org.springframework.security.core.GrantedAuthority
 import org.springframework.security.core.userdetails.UserDetails
-import org.springframework.stereotype.Component
 import java.util.UUID
 
-@Component
 class AuthDetails(
         private val userId: UUID
 ) : UserDetails {
