+++ notdo-infrastructure/src/main/kotlin/com/gsm/notdo/global/security/properties/JwtProperties.kt
@@ -1,15 +1,9 @@
 package com.gsm.notdo.global.security.properties
 
-import org.springframework.boot.context.properties.ConfigurationProperties
-import org.springframework.boot.context.properties.ConstructorBinding
-import java.util.Base64
-
-@ConstructorBinding
-@ConfigurationProperties(prefix = "jwt")
-class JwtProperties(
-        secretKey: String,
-        val accessExp: Int,
-        val refreshExp: Int
-) {
-    val secretKey = Base64.getEncoder().encodeToString(secretKey.toByteArray())
+object JwtProperties {
+    const val HEADER = "Authorization"
+    const val PREFIX = "Bearer "
+    const val ACCESS = "access"
+    const val REFRESH = "refresh"
+    const val AUTHORITY = "authority"
 }
